IHub-173327 Update the integration to work with the mock service
