+++ mock/README.md
@@ -1,7 +1,9 @@
 # Upsource mock service
 
+The Mock Service is for testing Upsource Integration.The Mock Service uses [Flask](https://flask.palletsprojects.com/en/1.1.x/) web framework.
+
 1. Install python Flask library (pip install Flask)
 
-2. Required to specify the location of the Flask application (export FLASK_APP=mock/mock_service.py)
+2. Specify the location of the Flask app (export FLASK_APP=mock/mock_service.py)
 
-3. To run a Flask app (flask run)
+3. Run Flask app (flask run)
